Add Implict Method Unit Tests
